client/requirements.txt
stellar-sdk>=10.0.0
python-dotenv
